move tools/transformers to transformers
