+++ tools/load_internlm_model.py
@@ -13,7 +13,7 @@ from internlm.initialize.launch import launch_from_torch
 from internlm.train import initialize_model
 from internlm.utils.registry import MODEL_INITIALIZER
 from internlm.utils.storage_manager import get_fns, init_storage_manager, llm_load
-from tools.transformers.interface import GenerationConfig
+from tools.interface import GenerationConfig
 
 logger = logging.getLogger(__file__)
 logging.basicConfig(level=logging.INFO)
